Renamed package get_item_by_id to get_clothing_item
