+++ src/main/java/interface_adapter/get_clothing_item/GetClothingItemPresenter.java
@@ -1,4 +1,4 @@
-package interface_adapter.get_item_by_id;
+package interface_adapter.get_clothing_item;
 
 import interface_adapter.ViewManagerModel;
 import use_case.get_clothing_item.GetClothingItemOutputBoundary;
