+++ src/main/java/interface_adapter/get_clothing_item/GetItemState.java
@@ -1,4 +1,4 @@
-package interface_adapter.get_item_by_id;
+package interface_adapter.get_clothing_item;
 
 import lombok.Getter;
 import model.ClothingItem;
